+++ handler/openid/flow_refresh_token.go
@@ -62,6 +62,9 @@ func (c *OpenIDConnectRefreshHandler) HandleTokenEndpointRequest(ctx context.Con
 	// We need to reset the expires at value
 	sess.IDTokenClaims().ExpiresAt = time.Time{}
 	sess.IDTokenClaims().Nonce = ""
+	sess.IDTokenClaims().JTI = ""
+	sess.IDTokenClaims().AccessTokenHash = ""
+	sess.IDTokenClaims().CodeHash = ""
 	return nil
 }